feat: cell layout view 파일 분리 #3
